refactors footer form into separate component
