+++ src/data/add_data.json
@@ -1,7 +1,7 @@
 {
   "app_data": {
     "about": "Community Health West London is a Community Interest Company made up of six local charities. We are working together with the wider community to improve the health and wellbeing of our residents.",
-    "disclamers": {
+    "disclaimers": {
       "copyright": "2021 All Rights Reserved by Community Health West London.",
       "accessability": "This site is built according to Web Content Accessibility Guidlines"
     }
